fixes + delete + deps

+++ src/test/resources/tl.cpp
@@ -7,8 +7,8 @@ int main() {
     ios_base::sync_with_stdio(0);
     long long x;
     cin >> x;
-    int mod = 1e9 + 7;
-    int res = 1;
+    long long mod = 1e9 + 7;
+    long long res = 1;
     for (long i = 1; i <= x; i++) {
         res = (res * i) % mod;
     }
